refactor: domain, view 패키지 추가
