+++ src/main/java/step3/domain/InvalidCarNameException.java
@@ -1,4 +1,4 @@
-package step3;
+package step3.domain;
 
 public class InvalidCarNameException extends RuntimeException {
     public InvalidCarNameException(String name) {
