+++ src/main/java/step3/view/ResultView.java
@@ -1,4 +1,8 @@
-package step3;
+package step3.view;
+
+import step3.domain.CarState;
+import step3.domain.JoinedCars;
+import step3.domain.RaceDisplay;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -11,9 +15,8 @@ public class ResultView {
         joinedCars.displayRaceStateOn(raceDisplay);
     }
 
-    public void showRaceWinner(List<Car> winners){
+    public void showRaceWinner(List<CarState> winners){
         String winnerNames = winners.stream()
-                .map(it -> CarState.of(it))
                 .map(it -> it.getName()).collect(Collectors.joining(","));
         raceDisplay.writeLine(winnerNames +"가 최종 우승했습니다.");
     }
