+++ src/test/java/step3/RaceRouletteTest.java
@@ -3,6 +3,7 @@ package step3;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import step3.domain.RaceRoulette;
 
 import static org.assertj.core.api.Assertions.assertThat;
