+++ src/test/java/step3/domain/JoinedCarsTest.java
@@ -1,0 +1,18 @@
+package step3.domain;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class JoinedCarsTest {
+
+    @Test
+    @DisplayName("구분자(,) 를 사용해서 자동차 이름을 생성자에 넘겨주면  ")
+    void create(){
+
+    }
+
+
+}
